Reorganize measures tests
